Add retrying mount manually on /opt/platform
There are two issues:
- When controller-1 is locked/unlocked with controller-0 removed,
  controller-1 does not start properly. No drbd related disks
  are mounted. The message "mount: can't find /opt/platform in
  /etc/fstab" in /var/log/daemon.log
- The error message "Device is held open by someone" in
  /var/log/damon.log
The first issue occurred because there is no entry for platform
n /opt/fstab as the message.
The second issue occurred because "drbdadm primary drbd-platform"
and "drbdadm secondary drbd-platform" are called one another.
(This happens when mount in platform fails)

This fix is:
- Add and call mount with parameter when mount commands fails
- Add sleep before "drbdadm secondary drbd-platform" is called

Closes-bug: 1966110

TEST PLAN:
PASS: Fresh install SX and DX
PASS: Path test for patch in SX
      1. Lock/Unlock controller-0
      2. Confirm controller-0 started properly and no error
         Addtional message is not shown in /var/log/daemon.log
PASS: Path test for patch in DX
      1. Delete controller-0
      2. Reboot controller-1
      3. Confirm controller-1 started properly and no error
         message is observed in /var/log/daemon.log

Signed-off-by: Takamasa Takenaka <takamasa.takenaka@windriver.com>
Change-Id: I238b9c13c9b66ebbc33207f7dba70cd0a45eca93