Cert-mon improvement
Currently, cert-mon is not aware of certificates created before
its startup. The kubernetes watch object monitors events coming after
the moment when it is called.

This change adds an initial listing of secrets in the 'watched'
namespaces and if there are any matches with secret_names in
CertWatcher objects, like RestApiCert_CertWatcher and
RegistryCert_CertWatcher, it will proceed to install the certificates.

This adds to cert-mon the ability of installing certificates created
before its startup.

Test plan:

PASS: Verify that a certificate created before cert-mon startup is
      successfully installed by cert-mon.
PASS: Verify that cert-mon does not re-install certificates every time
      it restarts.
PASS: Verify that cert-mon will re-install certificates in the event of
      certificate renewals.
PASS: Verify that cert-mon will also install certificates that are
      created after it is running.

Story: 2009811
Task: 44739

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I27ca17fc03c1439d6c0051b78a35436aec7c628b