Add Application Framework support for FluxCD
All system application-[upload/apply/remove/delete/abort/update]
need to be able to operate on both and Armada
tarball and a FluxCD tarball.
An app tarball will be considered in FluxCD format if it
contains a directory named "fluxcd-manifests".
And if the tarball does not have this structure
it will be checked further if it's in Armada format.
This first commit introduces full support for the
application-[upload/apply] on FluxCD tarballs, the other operations
are partially supported and improvements should be made
on upcoming reviews.

Tested upload/apply operations on the following apps:
- vault
- nginx
- auditd

Task: 44830
Story: 2009138
Change-Id: I7a34571de1f990e843a9d01375a9dd7732201c0c
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>