Fix lldp format in debian environment
Centos and Debian are using different versions of lldp tool.

Centos (python2) uses an older lldp 0.9.0
Debian (python3) uses a newer lldp 1.0.11

When calling the lldp tool passing 'json' flag, the results
are formatted differently in those versions of the tool.

A "json0" flag was added for backward compatability but
does not exist in 0.9.0.

For this solution:
'json' argument is kept for the older lldp on Centos.
'json0' argument is used for the newer lldp on Debian.

The results have some slight variations, so some special
handling in the processing methods is being added.

For Centos the results will be handled exactly the same.
For Debian, certain areas of the data need to be handled
differently.

Test Plan:
  PASS: Verify the KeyErrors are no longer generated in Debian
  PASS: Verify no new exceptions are generated in Debian.
  PASS: Verify Centos still works.
  PASS: Verify the controllerconfig changes can still be parsed

Closes-Bug: 1965978
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I133e934f58fcc03c2625ec818094acea23e21c12