Resize systemcontroller filesystems on upgrade-activate
On commits [1] and [2] the default filesystem sizes for backup
and platform are changed on System Controller role to increase
subcloud deployment paralellism, but these changes won't have
effect on an upgrade. So, a different mechanism is needed to
resize the filesystems in the upgrade scenario.

This commit adds an upgrade-script, to be executed during the
step upgrade-activate, to check the platform filesystem size
and, if below 20GB, resize platform fs to 20GB and backup fs
in 20GB - <current_platform_size>. If for any reason there is
no disk space to resize the filesystems this won't prevent the
upgrade to continue, but a warning message will be logged.

[1] https://review.opendev.org/c/starlingx/ansible-playbooks/+/835310
[2] https://review.opendev.org/c/starlingx/config/+/835311

Test Plan:
PASS: Upgrade Standard System Controller
      Verify platform and backup filesystems with increased size
      Verify no failures on upgrade process
PASS: Upgrade non-DC deployment
      Verify no change on filesystems size
      Verify no failures on upgrade process
PASS: Increase cgts-vg to fit resize
      Upgrade AIO DX System Controller
      Verify platform and backup filesystems with increased size
      Verify no failures on upgrade process
PASS: Upgrade a DC deployment without sufficient space to resize
      Observe no change on filesystems size
      Verify no failures on upgrade process

Story: 2009725
Task: 44892
Change-Id: I77eca98a5b2844baf23d7eab817ccc7da7ae8cf6
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>