Cleanup tox for python3.9 jobs
sysinv tox could not be run in a developer environment
due to a relative path that only existed in a zuul env.

the py39 zuul jobs do not need root repo cloned since
the upper constraints are accessed by URL.

The py files are only being updated in their header
so that the tox jobs will be triggered to be executed
due to zuul filtering.

Story: 2006796
Task: 44861
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I184dc910717239fd343a70c343361a4c6b8fb9e3