Adding unit tests for lldp agent driver activities
The lldp CLI returns different data on Debian than Centos
and so new unit tests are being written to validate the
code that processes those results

The unit tests are based on json data files that are
generated for those versions of the tool.

This submission covers the 0.9.0 version of the tool
which is currently the version for STX 6.0 on Centos

Current level of coverage is 70% for the lldp driver,
with the majority of the uncovered lines relating
to the input data not having vlans.

Partial-Bug: 1965978
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I8827e3652cdaa04e4c42dffaa85b92bee0e382f0