Add Sysinv enhancements for FluxCD
In this commit we added 2 enhancement to
the the FluxCD functionality:

1. Dynamically change the host in the default helm repository with the
system controller network address.

2. We will see this issue
https://github.com/fluxcd/helm-controller/issues/81
on AIO-SX if there are issues during chart install.
Basically, the status of helmrelease ends up with ready but
the pods are not actually ready/running.
This is due to helm upstream issues
https://github.com/helm/helm/issues/3173,
https://github.com/helm/helm/issues/5814,
https://github.com/helm/helm/issues/8660.
To solve this we need to check if the pods of the
applied helm charts are ready/running using
the kubernetes python client after the helmrelease is
in a ready state.

3. Check for the 'failed' state of the helmreleases and
update the app accordingly

4. Move the Timeout counter before starting the fluxcd
operations to prevent some infinite loops

Test Plan:
PASS: Deployed a SX with the 'cluster_host_subnet' changed from the
default one and checked if the helm repositories were different
as expected
PASS: Apply nginx fluxcd app 1.1-24 and verified that the app status
is 'applied' when all the pods are in running state
PASS: Apply vault fluxcd app 1.0-27 and verified that the app status
is 'applied'  when all the pods are in running state
PASS: Platform Upgrade from latest release to current release

Task: 44912
Story: 2009138
Change-Id: I207b5b55a4b504a1c8ecdb239036a3d122294a0d
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>