Exponential Backoff retry during image download
During a batch application apply on a large number of suclouds, the
massive image download may congest the MGMT network. This commit sets
up an exponential backoff logic to prevent the retry at the same time.

Test passed:
Apply a large application in parallel on a large number of subclouds.

Story: 2009725
Task: 45178
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: Ia7e5ab8fdcfcf148a0e769d0bea21ae577822143