Adjust agent to handle Broadcom NIC corner cases
While testing a load installed in a host with the Broadcom BCM57504 NIC
it was identified that interfaces were not displaying the correct speed
when queried through ethtool, reason being that for this specific NIC
they may take a little while to be properly operational. It was also
noticed that SR-IOV interfaces can only be set up after the
corresponding PF is up.

This change ensures that Broadcom interfaces are operational before
probing for their line rates. It also adds a flag to be passed to puppet
that ensures that VFs are set up only after the conrresponding PF is
administratively up. It adds 25Gbps as a valid port speed and ensures
that "ip link set up" and "ip link set down" commands are properly
waited for.

TESTS:
- Installed on a host with a BCM57504 NIC successfully
- Interface speed is displayed correctly when checked with ethtool
- VFs can be created successfully through system commands

Story: 2009915
Task: 45007
Signed-off-by: Antonio Augusto Vilas Boas Teixeira
<antonio.augustovilasboasteixeira@windriver.com>
Change-Id: I2519114a3deb9bd51a35ab1ded628df095f0dc1b