Fix log formatter errors in service file
Occasionally the log formatter reports errors on Debian.

The assumption is that during the signal handling,
the localization conversion returns a value without
parameters, and this causes the log formatter error
while using that invalid string along with lazy logging
for its parameters.

The localization code has been removed since it is
only meant to be used for customer facing interactions
like CLI.

The lazy logging was also removed. Lazy logging does
provide a performance improvement, but that optimization
would not be observed for these sparse logs.

Test Plan:
 PASS: Build/Install/Bootstrap/Unlock on Debian
 PASS: Build/Install/Bootstrap/Unlock on Centos
 PASS: Verify that there are no stacktraces related
to log formatting in sysinv.log in either OS.

Story: 2008943
Task: 45117
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I77376c98db7158dc18d5483974e0d5fbe9f94078