Block filesystem resizes if waiting for agent to report
Currently the system may allow multiple host-fs-modify commands
to be executed in sequence and in a short interval even if these
would consume more space than is available in the volume group.
This is caused because the volume group available space is updated
by the agent audit, on [1], which runs every minute, so every
fs resize command that is executed before the agent reports back
will be allowed to run and will update the fs size on the database,
leaving an inconsistent value, since when the puppet manifest runs
on the host it will fail with "not enough space" error (but won't
fail the manifest as return code 5 is allowed on [2], which is
the return code for this error, as can be observed on [3]).

The scenarios affected by the behavior described previously are:
host-fs-modify -> controllerfs-modify
host-fs-modify -> host-fs-modify

This commit adds an extra validation before allowing both "fs-modify"
commands to run, that compares the most recent host-fs "updated_at"
date with the cgts-vg "updated_at" date stored on the database, only
allowing the commands to run if the latter is greater than the former.

Test Plan:
PASS: run host-fs-modify command successfully and shortly after run
      another host-fs-modify command and verify that it is blocked
      from running
PASS: run host-fs-modify command successfully and shortly after run
      a controllerfs-modify command and verify that it is blocked
      from running
PASS: run host-fs-modify command successfully, wait for lvg available
      size to be updated and run another host-fs-modify, verify that
      it runs successfully
PASS: run host-fs-modify command successfully, wait for lvg available
      size to be updated and run a controllerfs-modify, verify that
      it runs successfully
PASS: resize host-fs and controllerfs before upgrade, upgrade subcloud
      and verify that upgrade runs successfully

Regression:
PASS: run host-fs-modify command that exceeds vg available space and
      verify that it is not allowed to execute
PASS: run controllerfs-modify command that exceeds vg available space
      and verify that it is not allowed to execute

[1] https://opendev.org/starlingx/config/src/branch/master/sysinv/sysinv/sysinv/sysinv/agent/manager.py#L1421
[2] https://opendev.org/starlingx/stx-puppet/src/branch/master/puppet-manifests/src/modules/platform/manifests/filesystem.pp#L146
[3] https://paste.opendev.org/show/bswm7v3Qj2cpiD2H9O3j/

Closes-bug: 1968201
Change-Id: Ifc2fb272f3762441cc0c7755443f01d9789f8723
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>