Add support for bonded interfaces configuration on Debian
This change corrects bonded interface configuration on Debian systems
using the ifupdown package.

The master interface now receives the parameter "bond-slaves" to
identify the physical interfaces, and the slaves receive the stanza
"alow-[master] [slave]" to trigger the configuration on the physical
interfaces.

During tests it was observed that the slaves only run the pre-up
step, so for them all commands are executed on that point since
slaves are the physical interfaces all pre-up can run since the
device exists.

Test Plan:
PASS configuration of balanced interfaces with all txhashpolicy
     values
PASS configuration of active-standby interfaces with all
     primary-reselect values
PASS configuration of addresses and routes on the bond interface
PASS configuration of VLAN interface on top of bond interface, with
     static addresses and routes
PASS all tests executed with host lock/unlock cycle

Story: 2009965
Task: 45196

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: Iec0d2b068c0a7a699aafd4af630364770fd3adb5