FluxCD: Fixes to support app auto re-applies
When the system is evaluating application overrides it needs to pass to
generate_helm_application_overrides() if the application is a FluxCD or
an Armada app.

Commit 8ab0a835e introduced changes for support the FluxCD application
framework and updated generate_helm_application_overrides(). This
change was missed.

In addition, when the application overrides are determined to have
changed and a auto-reapply is needed a flag file is generated to signal
that the app should be re-applied at the appropriate time.

This will also change the location of the generated flag files to the
Helm overrides base directory from the Armada base directory as this
directory will only exist after the first Armada application is
uploaded. Without this update, FluxCD apps are unable to auto re-apply
without an Armada app being present.

Test Plan:
 - Include change in a running AIO-DX CentOS system
 - Apply a FluxCD app
 - Lock/unlock controller-1 requiring override changes for replicas
 - Confirm no errors in sysinv log
 - Confirm the correct number of replicas for 1 and 2 unlocked
   controllers
 - Confirm auto apply
 - Perform the above with and Armada app present

Change-Id: Iee797a654a171e5b24ce3b04e503ffc8dc6f870b
Closes-Bug: #1970804
Signed-off-by: Robert Church <robert.church@windriver.com>