FluxCD: Allow 'Completed' pods when verifying app is ready
Due to some upstream errors specified in 60c75cf2, code was provided to
verify that apps which label their pods with app.kubernetes.io/name get
special treatment on AIO-SX to verify that they are running before
reporting that the application apply is complete.

This original code does not account for the pods that run jobs that go
to a 'Completed' state and cause the apply to be stuck and eventually
timeout.

This update will now check for pods in the 'Completed' state and will
also partially re-factor the code from 60c75cf2 to provide better
readability, maintainability, and increased logging.

Test Plan:
PASS - Build/Install AIO-SX + AIO-DX
PASS - Bootstrap and unlock AIO-SX + AIO-DX
PASS - Test FluxCD apps with the appropriate labeling and confirm
       running/completed conditions

Change-Id: I00fa35a2eef5f0d18def1a5233540503d7c1f212
Closes-Bug: #1971053
Signed-off-by: Robert Church <robert.church@windriver.com>