Prevent invalid configuration of max VF tx rate
If a port has its detected speed as None (e.g. port is not cabled), when
configuring the 'max-tx-rate' of an interface which uses this port any
value can be set, even one exceeding the NIC's supported speed.

This change adds an error message in this scenario, preventing the
configuration of the tx rate parameter when the port speed is unknown.

Closes-Bug: 1971575

Signed-off-by: Antonio Augusto Vilas Boas Teixeira
<antonio.augustovilasboasteixeira@windriver.com>
Change-Id: I271656df1bac03f9a8f714f876baf1c742748b31