Fix Mellanox device detection
There are some instances where the is_a_mellanox_device() method could
fail depending on how the driver of a port is detected. The driver
parameter of a port can be a string of various driver names, separated
by commas. The Mellanox detection method was not able to handle this and
would erroneously identify a device as a non Mellanox in this scenario.

TESTS:
- Installed on a host using a Mellanox CX5 card
- Set the interface using a port of the CX5 to pci-sriov class
- Confirmed that the port driver is listed as 'mlx5_core,mlx5_core' when
  running 'system host-port-show'
- Assigned this interface to a datanetwork
- Confirmed that the flag 'isRdma' is now correctly appended to the
  datanetwork resource in the file /etc/pcidp/config.json

Closes-Bug: 1973361

Signed-off-by: Antonio Augusto Vilas Boas Teixeira
<antonio.augustovilasboasteixeira@windriver.com>
Change-Id: Ib12d104a61c3e78803e4387804cd53a98ee8f0ec