[PTP] Generate hieradata to set currentUtcOffsetValid
A node running as a GM needs to be able to set the currentUtcOffsetValid
flag in ptp announce messages.

This change allows a user to set a ptp instance parameter:
currentUtcOffsetValid=1

This will put the required pmc command in the instance hieradata so that
the puppet manifest can configure the ptp4l process to set the flag.

Removing the currentUtcOffsetValid parameter and re-applying the ptp
config un-sets this flag.

A subsequent change to the ptpinstance puppet manifest will be submitted
to apply this setting.

Tests:

Pass: Build and install on AIO-SX
Pass: Validate that parameter is set and removed correctly, check
that the flag is present in Announce messages
Pass: Validate that the system TAI clock has the correct offset
from CLOCK_REALTIME

Story: 2010033
Task: 45379

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: Iaabf6188ddd047697476bf0dc1672c11a0e248a8