py3: Fix helm-override-update py3 compatibility
On Debian it is observed that 'system helm-override-update ...' fails.
Inspecting the logs we see an error related to a file write, that
expects a bytes type object instead of string type object.

Make code compatible with both PY2(CentOS) and PY3(Debian).

Tests on Debian and CentOS:
PASS: 'system helm-override-update ...' works

Story: 2006796
Task: 45454
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Idf4e4da76d6e495010cbea665e5dfc881ba85671