Do not add '::1' as a DNS nameserver
During bootstrap on Debian it was noticed that the file resolv.conf
is receiving 'nameserver ::1' along with the other DNS servers set
in localhost.yaml. This happens because the initial configuration
is being applied and the only address available for the host is
'::1'. Since ::1 is not a useful DNS address it might lead wrong
request execution during address resolution.

This change evaluates the gethostbyname response and if only contains
'::1' it will ignore it.

Test Plan:
PASS  Install IPv4 on Debian (AIO-SX)
PASS  Install IPv6 on Debian (AIO-SX)
PASS  Install IPv4 on CentOS (AIO-SX)
PASS  Install IPv6 on CentOS (AIO-SX)

Closes-Bug: 1973051

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: Ifb6f711fa832d7320872617cb9b508c156f2ae79