Adding nvidia_vgpu drivers support
On compute nodes with openstack-compute-node=enabled label,
the non-open-source NVIDIA vGPU drivers should, if present,
be loaded for commercial scenarios where the non-open-source
drivers are built into the StarlingX ISO.

This change adds the conditional set of a config flag that
controls the non-open-source vGPU drivers.

Test plan:
PASS: Verify that after applying the modification in a host
WITH openstack-compute-node=enabled label tagged the installation
script of nvidia_vgpu drivers (if present) should be triggered
to be executed.
PASS: Verify that after applying the modification in a host
WITHOUT the openstack-compute-node=enabled label tagged the
uninstall script of nvidia_vgpu drivers (if present) should be
triggered to be executed.

Story: 2010090
Task: 45618

Signed-off-by: Rafael Falcao <rafael.vieirafalcao@windriver.com>
Co-Authored-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: I4d22398c551a3462cad0d040e8955365dda3dc99