Fix tox-docs failing sphinx
A new version of sphinx was released May 29 2022
which requires a language setting in config otherwise
a warning (treated as error) causes the sphinx operation
to fail.

Updating a file header to trigger zuul.
Updated the sphinx config file to correct the issue.

The sphinx behavioural change is mentioned here:
https://github.com/sphinx-doc/sphinx/issues/10062
https://github.com/sphinx-doc/sphinx/issues/10474

Closes-Bug: #1976377
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I4d489a7bb8f579c894081bedba8ac3638a7eae1d