Config support for nohz_full on standard workers
This commit ensures the nohz_full configuration on standard workers by
default. If there is a need to keep the existing behavior on standard
workers a label must be added to it in order to disable the nohz_full
config.

Story: 2010062
Task: 45514

Test Plan:
PASS: Bootstrap a standard worker with starlingx master and check if
      parameter was set.
PASS: Bootstrap a standard worker with label configured in dm config
      file.
PASS: Remove label and check if the parameter was unset.

Signed-off-by: Iago Estrela <IagoFilipe.EstrelaBarros@windriver.com>
Change-Id: I53f9863dd7c13d3f91e8cd1caed21a8646a5da89