FluxCD: Support disabling release instead of chart
The initial implementation of the FluxCDKustomizeOperator was following
the ArmadaManifestOperator wrt disabling charts of a given app, but
FluxCD has this layer of abstraction called HelmRelease. Most of the
StarlingX apps are defining the HelmRelease with the same name of the
charts and we were disabling the releases all the way through the
initial implementation anyway, but if the HelmRelease have a different
name of the chart this will fail. Besides that, we might want to define
more than one release using the same chart. This fix will allow
fulfilling both cases.

Also, this change implements a new sysinv.helm.base.FluxCDBaseHelm that
we encourage to be extended for the helm plugins of apps that are using
FluxCD so they require CHART and HELM_RELEASE to be defined.

It's worth mentioning that I had to change the way some mocks were done
on test_kubernetes.py and test_manager.py since they were producing
false positives on kustomize_base.py#L139.

TEST PLAN
PASS Upload app with charts disabled on the FluxCDKustomizeOperator,
     check if they are not present on the list of resources on the
     top-level kustomization.yaml

Story: 2009138
Task: 45572

Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I96bf49601d1227cb876d7f9e54d6efd799a56a53