Track config applied to align with reboot required on first report
The runtime config update is generated on the active controller for
a runtime config apply. However, it is valid and may also be applied
as part of the host-unlock/initial report.

Configuration updates are checked for match of reboot required config
if it is as part of a first report from the sysinv agent.

TestPlan:
PASSED perform subcloud add, bootstrap and manage
PASSED apply DX runtime config which inherits reboot apply host-unlock
PASSED Perform host-lock/unlock AIO Duplex

Closes-Bug: 1977852
Signed-off-by: John Kung <john.kung@windriver.com>
Change-Id: I7a0404a9daa521a8886a1a871ae7160bae3fdc72