Handle status_code / status error in HTTP response
Checks for the existence of either 'status_code'
'status_int', or 'status' in the HTTP response object,
instantiating the HTTPException instance with the found
code, or raises a generic Exception if not found.

Test Plan:

PASS:
Tested code on system exhibiting the errors
referenced in the bug. The HTTP response is now
correctly caught and the correct HTTPException is
raised.

Closes-Bug: 1978499

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I53d1dd6583368de4f6713a838fe3304d362f1756