Delete certificate alarm when secret is deleted
For certificates stored as kubernetes tls secrets, the alarm should be
cleared when the secret is deleted.

This changes the audit_for_deleted_certificates function to also check
for deleted secrets and subsequently clear the alarm and delete the
certificate snapshot information.

Test plan:

PASS: Verify that when a certificate is deleted the alarm is cleared
      from the system
PASS: Verify that deploying a soon-to-expire certificates results in
      an alarm in 'fm alarm-list'
PASS: Verify that an existing certificate alarm is cleared up by
      renewing the certificate to get a valid certificate

Closes-Bug: 1978730

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I6ed9248766b2abbbcc616e10d4575b4ae0471c9d