system health-query response on ceph query
In order to assure a response to the system health-query, when Ceph
storage-backend is configured and the ceph-api is unresponsive,
a Timeout is required.

This Timeout does not rely on the underlying ceph-api timeout as
the ceph-api may not timeout as expected.

Test Plan:
PASSED Verify system health-query response when Ceph is unhealthy
PASSED Verify system health-query response when Ceph is healthy

Closes-Bug: 1978726
Signed-off-by: John Kung <john.kung@windriver.com>
Change-Id: I4702c409e8ea45946ba94fab6a0989a90f2f6604