[PTP] Update notification namespace to be privileged
The ptp-notification application requires a privileged namespace in
order to deploy and operate.

This change moves the notification namespace from the baseline policy
group to the privileged policy group so that it can continue to operate
as it did prior to the addition of support for the Pod Security
Admission controller introduced in the upversion to k8s 1.23.

The privileged and baseline groups were defined in
https://review.opendev.org/c/starlingx/config/+/833487

Test-plan:

Pass: Update the privileged and baseline groups in common.py, restart
sysinv-conductor and verify that ptp-notification is able to properly
deploy.

Pass: Verify that the notification namespace has the expected
privileged labels.

Closes-Bug: 1978737

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I5d24a8e81b32809f568a5953701cf2e0c474005e