Disallow without vlan id when creating vlan interface
When configuring an interface to vlan, it is possible to configure the
interface without entering a vlan id. If vlan id is set without input, vlan id
is set to NONE. If this interface is assigned to oam and unlocked, a problem
occurs. After unlock, the device name is set to vlanNone and the oam interface
cannot up normally.

Closes-Bug: 1979253

Test Plan:
PASS: Create vlan interface with vlan id and succeed in creation
PASS: Create vlan interface without vlan id and fail in creation

Signed-off-by: ohjoohyun <oh011798@gmail.com>
Change-Id: Ifd5527699cb8e8f874a54292f40656fc890f0a1a