sysinv-api: Temporarily increase timeout to let bootstrap pass
On Debian we observe that bootstrap fails intermittently.
From the data so far, this is observed only on VMs.

Couldn't reproduce the error locally, but from the investigation
deduce it is just a performance degradation.
Investigation notes to prove it is a performance degradation and not a
service crash are uploaded to the LP.
Temporarily increase the retries for waiting for sysinv-api to come up.
Jump from 15 seconds to 60 seconds to be defensive.
This will allow sanities to pass and integration effort to continue.

Tests on AIO-SX on Debian:
PASS: bootstrap

Partial-Bug: 1979717
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I10b00eab467303771d25cc5c79760005c3966446