Use session object when creating keystone_client
The new version of the keystone client we are using in debian does not
support instantiating a keystone_client without a session object.

This is making use of an existing self._get_keystone_session function
to create the session object and using that session object to
authenticate with the keystone_client. There are also minor
refactorings to avoid using the keystone client at bootstrap time.

Test plan on both centos and debian:

PASS: Run 'openstack user password set' to change the password and
      verified that /opt/platform/puppet/22.06/hieradata/
      secure_system.yaml gets updated with the new password, keyring
      gets updated with new password and /var/log/sm-customer.log
      shows that the vim service gets restarted

PASS: Verify that no keystone error messages are shown during bootstrap
      in sysinv.log keystone listener are started.
      
PASS: When keystone admin user password changes, observe the 
      corresponding entity in keyring is updated, and vim is restarted
      (by openstack::keystone::password::runtime)

Closes-Bug: 1979995

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: Id4d4bf8072a853d4d3e016e9afa2cd984ee85694