Increase the audit interval of k8s local secrets to once per day
This audit only purpose is to ensure synchronization between k8s
secrets and keyring, but there's already a notification system by
keystone that sync this data.

After a research, we noticed that the appropriate time interval
for this audit is to be executed only once per day.

This commit is a continuation of the study started at task 45616.

TEST PLAN:
PASS: AIO-SX: manually replaced these files into a Debian installation
and no crashes happened.
PASS: rebuild the whole system with the modification with no crashes.
PASS: AIO-SX: verify that the audit is being called during the period.
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: guarantee that all audits are being called by the periodic tasks.
PASS: perform a host-lock/unlock and verify that the values at
sysinv.conf remains the same.
PASS: change the keystone admin password via cli and verify if the
keyring also actualizes the password

Story: 2010087
Task: 45719

Depends-On: https://review.opendev.org/c/starlingx/config/+/848437
Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: I572b093eed99c9f1616510ca2d4788cf354f20ef