Ignore JSON decode error in request log.
For DELETE requests with no request body or "Content-Length" field, the
expected return value for "hasattr(rest_state.request, 'json')" would be
False and for "rest_state.request.json" would be an empty string.
Instead, what occurs is that both "hasattr(rest_state.request, 'json')"
and "rest_state.request.json" are throwing JSON decode exceptions in
this case. This change adds tolerance for this exception. This is done
because it is not know all the possible request conditions that may
cause this exception to be thrown.

Test Plan:

PASS: Successfully deploy an AIO-SX using a Debian image with this
commit present and, using horizon interface, create and delete a data
network, checking that in log file "/var/log/sysinv-api.log" no
exception was thrown in the delete request.

Closes-Bug: 1980842
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: Ib9f72f3b2d4d60790fa8011674a013bcd141a61e